requirements.txt
# ─────────────────────────────────────────────────────────────
# MR AI RAG v2 — Python Dependencies
# Python 3.11+ recommended
# ─────────────────────────────────────────────────────────────

# ── Web Framework ─────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
aiofiles==24.1.0

# ── Data Validation ───────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.7.0

# ── PDF Parsing ───────────────────────────────────────────────
PyMuPDF==1.25.1

# ── Local Embeddings (no API key needed) ──────────────────────
sentence-transformers==3.3.1

# ── Vector Store ──────────────────────────────────────────────
faiss-cpu==1.9.0.post1

# ── LLM Providers ─────────────────────────────────────────────
openai==1.58.1
anthropic==0.40.0
google-generativeai==0.8.3
httpx==0.28.1

# ── YouTube & Video Transcription ─────────────────────────────
yt-dlp>=2024.1.1
youtube-transcript-api>=0.6.1
openai-whisper>=20231117
# NOTE: ffmpeg installed via apt in Dockerfile

# ── Utilities ─────────────────────────────────────────────────
numpy==1.26.4
python-dotenv==1.0.1